% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{sortImages}
\alias{sortImages}
\title{Sort annotated images}
\usage{
sortImages(ann_fn, coord_fn = NULL, out_dir = "")
}
\arguments{
\item{ann_fn}{A path to an annotation file}

\item{coord_fn}{A path to a coordinate file}

\item{out_dir}{A path to an output directory}
}
\description{
Sort annotated images
}
