% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{sliceImages}
\alias{sliceImages}
\title{Slice an image}
\usage{
sliceImages(
  img_fn,
  width = 100,
  height = 100,
  step = 80,
  divide = 0,
  fmt = "auto",
  out_dir = "",
  out = "slice"
)
}
\arguments{
\item{img_fn}{A path to an image file}

\item{width}{An integer to specify the width of output sliced images}

\item{height}{An integer to specify the height of output sliced images}

\item{step}{An integer to specify the sliding window step size}

\item{divide}{An integer to specify how many tiles the image is cut into.}

\item{fmt}{A string to specify the output file format}

\item{out_dir}{A path to the output directory}

\item{out}{The prefix of output files}
}
\description{
Slice an image
}
